There are now more than 32 different coda interfaces, thus the codadebug variable needs more than 32 bits. Thus change turns codadebug into a long long giving 64 bits to play with.

+++ sys/darwin-7.3/coda/cnode.h
@@ -101,7 +101,7 @@ extern int coda_nc_use;
 /*
  * Used to select debugging statements throughout the cfs code.
  */
-extern int codadebug;
+extern long long codadebug;
 extern int coda_nc_debug;
 extern int coda_printf_delay;
 extern int coda_vnop_print_entry;
